==> code/01 Naver Stock Price.ipynb <==
## Naver Stock Price

- 아래 링크로 조회되는 주식 가격을 수집합니다.(삼성전자)
- 링크: https://finance.naver.com/item/main.naver?code=005930

### HTTP 요청 실행

# requests 라이브러리를 설치합니다.
!pip install requests

# 관련 라이브러리를 호출합니다.
import requests

# 상장회사의 종목코드를 설정합니다.(삼성전자)
stock = '005930'

# 요청 URL을 설정합니다.
url = f'https://finance.naver.com/item/main.nhn?code={stock}'

# HTTP 요청을 실행합니다.
res = requests.get(url = url)

### HTTP 응답 확인

# HTTP 응답 상태코드를 확인합니다.
res.status_code

# HTTP 응답 헤더를 확인합니다.
res.headers

# HTTP 요청 URL을 확인합니다.
res.url

# HTTP 응답 바디(str)를 출력합니다.
res.text

### HTML 파싱(Parsing)

# bs4 라이브러리를 설치합니다.
!pip install bs4

# 관련 라이브러리를 호출합니다.
from bs4 import BeautifulSoup as bts

# 문자열(str)을 bs4.BeautifulSoup 자료형으로 변환합니다.
soup = bts(markup = res.text, features = 'html.parser')

# soup의 클래스를 확인합니다.
type(soup)

# soup을 출력합니다.
soup

### 현재 주식 가격 수집

# 현재 주가 정보를 포함하는 HTML 요소를 선택하고 items에 할당합니다.
items = soup.select(selector = 'div.today > p.no_today > em > span')

# items의 길이를 확인합니다.
len(items)

# items을 출력합니다. [참고] 첫 번째 원소에 완성된 텍스트가 포함되어 있습니다.
items

# items의 첫 번째 원소를 출력합니다.
items[0]

# 위 코드 실행 결과에서 텍스트를 추출합니다.
items[0].text

# 현재 주가를 price에 할당합니다.
price = items[0].text
print(price)

# price를 정수로 변환하기 전에 콤마를 삭제해야 합니다.
price = int(price.replace(',', ''))
print(price)

### 전일 대비 수집

# 전일 대비 정보를 포함하는 HTML 요소를 선택하고 items에 할당합니다.
items = soup.select(selector = 'div.today > p.no_exday > em > span')

# items을 출력합니다. [참고] 두 번째 원소에 완성된 텍스트가 포함되어 있습니다.
items

# items의 두 번째 원소를 출력합니다.
items[1]

# 위 코드 실행 결과에서 텍스트를 추출합니다.
items[1].text

# 전일 대비를 change에 할당합니다.
change = items[1].text
print(change)

# change에서 콤마를 삭제하고 정수형으로 변환합니다.
change = int(change.replace(',', ''))
print(change)

## End of Document
