==> code/02 Naver Market Index.ipynb <==
## Naver Market Index

- 아래 링크로 조회되는 환율, 금, 휘발유 등 전일자 주요 시장지표를 수집합니다.
- 링크: https://finance.naver.com/marketindex/

### HTTP 요청 실행

# 관련 라이브러리를 호출합니다.
import requests

# 요청 URL을 설정합니다.
url = 'https://finance.naver.com/marketindex/'

# HTTP 요청을 실행합니다.
res = requests.get(url = url)

### HTTP 응답 확인

# HTTP 응답 상태코드를 확인합니다.
res.status_code

# HTTP 응답 헤더를 확인합니다.
res.headers

# HTTP 요청 URL을 확인합니다.
res.url

# HTTP 응답 바디(str)를 출력합니다.
res.text

### HTML 파싱(Parsing)

# 관련 라이브러리를 호출합니다.
from bs4 import BeautifulSoup as bts

# 문자열(str)을 bs4.BeautifulSoup 자료형으로 변환합니다.
soup = bts(markup = res.text, features = 'html.parser')

# soup의 클래스를 확인합니다.
type(soup)

# soup을 출력합니다.
soup

### 환전 고시 환율 수집

# 환전 고시 환율 정보를 포함하는 HTML 요소를 선택하고 items에 할당합니다.
items = soup.select(selector = 'ul#exchangeList > li')

# items의 길이를 확인합니다.
len(items)

# items을 출력합니다. [참고] 네 개의 환율 관련 HTML 요소가 반복되어 있습니다.
items

# items의 첫 번째 원소를 출력합니다.
items[0]

# items의 첫 번째 원소를 item에 할당합니다.
item = items[0]

# item에서 환전 고시 환율 통화명을 포함하는 HTML 요소를 선택합니다.(리스트)
item.select(selector = 'span.value')

# 위 코드 실행 결과에서 첫 번째 원소를 선택합니다.(bs3.element.Tag)
item.select(selector = 'span.value')[0]

# 위 코드 실행 결과에서 텍스트를 추출합니다.
item.select(selector = 'span.value')[0].text

### for 반복문 실행

# 지금까지의 과정을 for 반복문으로 실행합니다.
for item in items:
    print(item.select(selector = 'span.value')[0].text)

# for 반복문 대신 리스트 컴프리헨션을 사용하여 리스트로 반환합니다.
[item.select(selector = 'span.value')[0].text for item in items]

### HTML 요소의 텍스트를 리스트로 저장하는 함수 생성

# HTML 요소의 텍스트를 반환하는 사용자 정의 함수를 생성합니다.
def getHtmlText(x, css):
    return [i.select(selector = css)[0].text for i in x]

# 환전 고시 환율을 리스트로 생성합니다.
rates = getHtmlText(x = items, css = 'span.value')
print(rates)

# 문자열에서 콤마를 제거하고 실수로 변환합니다.
rates = [float(rate.replace(',', '')) for rate in rates]
print(rates)

### 유가, 금시세 수집

# 유가, 금시세를 포함하는 HTML 요소를 선택하고 items에 할당합니다.
items = soup.select(selector = 'ul#oilGoldList > li')

# items의 길이를 확인합니다.
len(items)

# 유가, 금시세를 리스트로 생성합니다.
prices = getHtmlText(x = items, css = 'span.value')
print(prices)

# 문자열에서 콤마를 제거하고, 실수로 변환합니다.
prices = [float(price.replace(',', '')) for price in prices]
print(prices)

## End of Document
