==> code/05 Naver Blog Link.ipynb <==
## 네이버 블로그에서 검색어로 링크 수집 (1 페이지)

### HTTP 요청 실행

# 관련 라이브러리를 호출합니다.
import requests

# 요청 URL을 설정합니다.
url = 'https://section.blog.naver.com/ajax/SearchList.naver'

# 검색어를 설정합니다.
searchWord = '스탠바이미'

# 오늘 날짜를 'yyyy-mm-dd'로 지정합니다.
today = '2022-04-04'

# 쿼리 스트링을 딕셔너리로 설정합니다.
query = {
    'countPerPage': 7,
    'currentPage': 1,
    'keyword': searchWord,
    'startDate': today, 
    'endDate': today, 
    'orderBy': 'sim',
    'type': 'post'
}

# 요청 헤더를 딕셔너리로 설정합니다.
headers = {
    'content-type': 'application/json;charset=UTF-8',
    'referer': 'https://section.blog.naver.com/Search/Post.naver',
    'user-agent': 'Mozilla/5.0 Chrome/99.0.4844.83 Safari/537.36'
}

# HTTP 요청을 실행합니다. 
res = requests.get(url = url, params = query, headers = headers)

### HTTP 응답 확인

# HTTP 응답 상태코드를 확인합니다.
res.status_code

# HTTP 응답 헤더를 확인합니다. [참고] Content-Type이 'json'입니다.
res.headers

# HTTP 요청 URL을 확인합니다.
res.url

# HTTP 응답 바디(str)의 일부를 출력합니다.
res.text[:1000]

### JSON 데이터 처리

# 관련 라이브러리를 호출합니다.
import json

# 불필요한 문자열을 삭제합니다. 
res.text.replace(')]}\',', '')

# HTTP 응답 바디를 딕셔너리로 변환합니다. 
data = json.loads(s = res.text.replace(')]}\',', ''))
print(data)

# data의 클래스를 확인합니다.
type(data)

# data의 key를 확인합니다.
data.keys()

# data['result']의 클래스를 확인합니다.
type(data['result'])

# data['result']의 key를 확인합니다.
data['result'].keys()

# data['result']에서 검색 결과 개수를 확인합니다.
data['result']['totalCount']

# data['result']에서 검색 결과 데이터를 선택합니다.
data['result']['searchList']

### 데이터프레임으로 변환

# 관련 라이브러리를 호출합니다.
import pandas as pd

# 위 코드 실행 결과를 데이터프레임으로 변환합니다.
df = pd.DataFrame(data = data['result']['searchList'])

# df의 처음 5행을 출력합니다.
df.head()

# 필요한 열이름만 리스트로 설정합니다.
cols = ['blogId', 'logNo', 'blogName', 'nickName', 'noTagTitle', 'addDate']

# 필요한 열이름만 남깁니다.
df = df[cols]

# df의 정보를 확인합니다.
df.info()

# logNo를 문자열로 변환합니다.
df['logNo'] = df['logNo'].astype('str')

# addDate를 날짜 데이터로 변환합니다.
df['addDate'] = pd.to_datetime(arg = df['addDate'], unit = 'ms')

# df의 처음 5행을 출력합니다.
df.head()

## End of Document
