==> code/08 Naver Cafe Link.ipynb <==
## 네이버 카페에서 검색어로 링크 수집 (1 페이지)

### HTTP 요청 실행

# 관련 라이브러리를 호출합니다.
import requests

# 요청 URL을 설정합니다.
url = 'https://apis.naver.com/cafe-home-web/cafe-home/v1/search/articles'

# 검색어를 설정합니다.
searchWord = '스탠바이미'

# 오늘 날짜를 'yyyymmdd'로 지정합니다.
today = '20220404'

# POST 바디를 딕셔너리로 설정합니다.
body = {
    'exceptMarketArticle': 1,
    'page': 1,
    'period': [today, today], 
    'query': searchWord,
    'sortBy': 0
}

# 요청 헤더를 딕셔너리로 설정합니다.
headers = {
    'content-type': 'application/json;charset=UTF-8',
    'referer': 'https://section.cafe.naver.com/ca-fe/home/search/articles',
    'user-agent': 'Mozilla/5.0 Chrome/99.0.4844.83 Safari/537.36'
}

# HTTP 요청을 실행합니다. 
res = requests.post(url = url, json = body, headers = headers)

### HTTP 응답 확인

# HTTP 응답 상태코드를 확인합니다.
res.status_code

# HTTP 응답 헤더를 확인합니다. [참고] Content-Type이 'json'입니다.
res.headers

# HTTP 요청 URL을 확인합니다.
res.url

# HTTP 응답 바디(str)의 일부를 출력합니다.
res.text[:1000]

### JSON 데이터 처리

# 관련 라이브러리를 호출합니다.
import json

# HTTP 응답 바디를 딕셔너리로 변환합니다. 
data = json.loads(s = res.text)
print(data)

# data의 클래스를 확인합니다.
type(data)

# data의 key를 확인합니다.
data.keys()

# data['message']의 클래스를 확인합니다.
type(data['message'])

# data['message']의 key를 확인합니다.
data['message'].keys()

# data['message']['result']의 클래스를 확인합니다.
type(data['message']['result'])

# data['message']['result']의 key를 확인합니다.
data['message']['result'].keys()

# data['message']['result']에서 검색 결과 개수를 확인합니다.
data['message']['result']['totalCount']

# data['message']['result']에서 검색 결과 데이터를 선택합니다.
data['message']['result']['searchResultArticle']['searchResult']

### 데이터프레임으로 변환

# 관련 라이브러리를 호출합니다.
import pandas as pd

# 위 코드 실행 결과를 데이터프레임으로 변환합니다.
df = pd.DataFrame(data = data['message']['result']['searchResultArticle']['searchResult'])

# df의 처음 5행을 출력합니다.
df.head()

# 필요한 열이름만 리스트로 설정합니다.
cols = ['articleid', 'clubid', 'clubname', 'rawWriteDate', 'subject']

# 필요한 열이름만 남깁니다.
df = df[cols]

# df의 정보를 확인합니다.
df.info()

# articleid와 clubid를 문자열로 변환합니다.
df = df.astype({'articleid': 'str', 'clubid': 'str'})

# rawWriteDate를 날짜 데이터로 변환합니다.
df['rawWriteDate'] = pd.to_datetime(
    arg = df['rawWriteDate'], 
    format = '%Y%m%d%H%M'
)

# df의 처음 5행을 출력합니다.
df.head()

## End of Document
